+++ src/Classes/Export.php
@@ -376,12 +376,12 @@ class Export extends \Backend
 				{
 					case '1': // Nur Mitglieder (Aktiver Mitgliedschaftszeitraum)
 						// Mitgliedschaften prüfen (memberships)
-						$exportieren = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($records->memberships);
+						$exportieren = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($records);
 						break;
 
 					case '8': // Alle Nichtmitglieder
 						// Nichtmitgliedschaften prüfen (memberships)
-						$exportieren = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($records->memberships);
+						$exportieren = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($records);
 						// Wahr/Falsch umdrehen
 						if($exportieren) $exportieren = false;
 						else $exportieren = true;