+++ src/Cron/Mitgliederpruefung.php
@@ -71,7 +71,7 @@ class Mitgliederpruefung
 					while($objPlayer->next())
 					{
 						// Ist der Spieler Mitglied im BdF?
-						$mitglied = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($objPlayer->memberships, NULL, $objPlayer->published);
+						$mitglied = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($objPlayer, NULL, $objPlayer->published);
 
 						if($mitglied)
 						{