+++ src/Cron/Mitgliedschaftsende.php
@@ -42,7 +42,7 @@ class Mitgliedschaftsende
 			while($objPlayer->next())
 			{
 				// Ist der Spieler noch Mitglied im BdF?
-				$mitglied = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($objPlayer->memberships, NULL, $objPlayer->published);
+				$mitglied = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($objPlayer, NULL, $objPlayer->published);
 				if(!$mitglied)
 				{
 					// Archivierung des Spielers notwendig, da nicht mehr Mitglied